+++ src/test/scala/edu/ie3/simona/agent/grid/congestion/data/CongestionManagementDataSpec.scala
@@ -7,9 +7,10 @@
 package edu.ie3.simona.agent.grid.congestion.data
 
 import edu.ie3.datamodel.models.result.CongestionResult.InputModelType
-import edu.ie3.simona.agent.grid.GridAgentData.GridAgentBaseData
+import edu.ie3.simona.agent.grid.data.GridAgentData.GridAgentBaseData
 import edu.ie3.simona.agent.grid.GridEnvironment
 import edu.ie3.simona.agent.grid.congestion.{CongestedComponents, Congestions}
+import edu.ie3.simona.agent.grid.data.CongestionManagementData
 import edu.ie3.simona.event.ResultEvent.PowerFlowResultEvent
 import edu.ie3.simona.test.common.UnitSpec
 import edu.ie3.simona.test.common.result.CongestedComponentsTestData